requirements.txt
google-genai>=0.8.0
Pillow>=10.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
cairosvg>=2.7.0
numpy>=1.26.0
